feat(PIO-49): set up tools monorepo workspace and move FlashViewClient to flashview-crypto
Add npm workspace root at tools/package.json linking flashview-crypto and
flashview-cli. Move FlashViewClient + ApiError into flashview-crypto/src/api.js
so both the CLI and upcoming mobile app can share the API client without
duplicating code. Add Retry-After header capture to ApiError for better 429
handling. The CLI's api.js becomes a thin re-export. All 105 CLI tests pass.
