+++ tools/package.json
@@ -1,0 +1,7 @@
+{
+    "private": true,
+    "workspaces": [
+        "flashview-crypto",
+        "flashview-cli"
+    ]
+}
